+++ src/core/three/types.d.ts
@@ -1,0 +1,7 @@
+import { ThreeInstance } from '@/lib/three-next';
+
+interface TestInstance extends ThreeInstance {
+  setCameraPosition: (x: number, y: number, z: number) => void;
+}
+
+export type { TestInstance };